+++ app/connectors/signMessagePage.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
 import * as sel from "../selectors";
-import { signMessageAttempt, signMessageCleanStore } from "../actions/SignMessageActions";
+import { signMessageAttempt, signMessageCleanStore } from "../actions/ControlActions";
 import { selectorMap } from "../fp";
 
 const mapStateToProps = selectorMap({
