+++ app/connectors/verifyMessagePage.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
 import * as sel from "../selectors";
-import { verifyMessageAttempt, verifyMessageCleanStore } from "../actions/VerifyMessageActions";
+import { verifyMessageAttempt, verifyMessageCleanStore } from "../actions/ControlActions";
 import { getMessageVerificationServiceAttempt } from "../actions/ClientActions";
 import { selectorMap } from "../fp";
